useNavigate running into issues
