--- src/client/login.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>BooMonkeyBoo Login</title>
-</head>
-
-<body>
-    <div id="log"></div>
-
-    <script src="login.bundle.js"></script>
-</body>
-
-</html>
